packages/monarch-mcp/pyproject.toml
[project]
name = "monarch-mcp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "mcp",
    "fastmcp>=2.0.0",
    "aiohttp",
    "pydantic",
    "httpx",
    "pytest",
    "pytest-asyncio",
    "openai",
    "python-dotenv",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
